+++ Content/Source/Data/Classes/UpgradeInfo.cs
@@ -11,8 +11,8 @@ public partial class UpgradeInfo : GodotObject
     public Array<string> Prerequisites { get; set; }
 
     // name : amount
-    public Dictionary<string, int> ResourceCost { get; set; }
+    public Dictionary<string, double> ResourceCost { get; set; }
 
     // name : paramater : isMultiplicative : effectValue
-    public Dictionary<string, Dictionary<string, Dictionary<bool, float>>> UpgradeEffects { get; set; }
+    public Dictionary<string, Dictionary<string, Dictionary<bool, double>>> UpgradeEffects { get; set; }
 }
